app: reimplement the process server
The implementation in orocos.rb is "ossified". It's supposed
to be a public API, to be compatible with Orocos.run (hint: it's
probably not) and because it's in orocos.rb, it can't be made
syskit-specific.

Reimplement the server and client parts within Syskit, marking
them as privaate implementations. I wanted to get rid of the
manual socket-handling, and use instead one of the existing event-
driven frameworks. Looked for EventMachine, but it turned out
that it is really hard to integrate it within another framework.

async-io seems to be in wide enough use (through the Falcon server)
and provide all we need, including an ease of integration within
another event loop, which would allow us to ultimately use it for
Syskit and Roby internal servers and clients as well (e.g. the
interface server)

It paves the path towards using proper TLS connections between Syskit
and its process servers, something that is very much needed.