implement aggregating multiple ports in data readers
The current dynamic port/reader binding implementation would always
pick the first matching port. This is fine in most cases, but in
*some* cases, one would like to aggregate data from "all components
matching X"

This commit adds the `all` option to data_reader for this purpose, as
well as the `with_task` option to transform.

The former instructs Syskit to read and return data from all matching
ports, i.e. with

   data_reader Some::Model.match.running.out_port, all: true, as: "my"

`my_reader.read` would return an un-ordered array of values. The
`with_task` option to `transform` allows to associate them with
the matching ports:

   my_reader.transform(with_task: true) do |port_to_value_map|
      # do some processing
   end